[FIX] In Portrait&Landscape mode video frame rePosition itself to closest corner of screen.
